poller/pyproject.toml
[tool.poetry]
name = "poller"
version = "0.1.0"
description = ""
authors = ["Tomi Kasurinen <tomi.kasurinen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.2"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
types-requests = "^2.28.11.14"
black = "^23.1.0"
mypy = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
